Moved ViewManagerGateway to correct package to eliminate incorrect dependency
